Update Makefile, FuzzBALL stuff, and README for build
Fix the Pin tool library linking arguments, which turned out to be the
last thing keeping it from running.
Get cfg_fuzzball compiling again, and run-fuzzball.pl running.
However for a variety of reasons the results do not yet reproduce
what had in the 2011 paper.
Start some instructions in the README file.