Remove references to obsolete cast types
This is for compatibility with FuzzBALL trunk commit d3b97833. It
correpsonds to r1744 from the CGC SVN branch, but it is needed with
trunk FuzzBALL too. Moving these build fixes earlier in the history
should make the CGC branch history compilable with a single FuzzBALL
version (roughly the CQE-time effc0b44).