First work on the static disassembly make-cfg tool
------------------------------------------------------------------------
r386 | smccamant | 2014-06-27 01:03:17 -0500 (Fri, 27 Jun 2014) | 7 lines

Start filling in the basics of a "make-cfg" tool which avoids
the need for Pin and only uses libelf and recursive disassembly.

So far: find the .text section with libelf, and generalize the old
disassembly code so that the address of the code bytes doesn't have
to equal the address they'll have one executed.