Add --addresses-file= option to make-cfg
Specifying the location of what was formerly always addresses.txt

------------------------------------------------------------------------
r1860 | jbenton | 2015-02-03 18:44:55 -0600 (Tue, 03 Feb 2015) | 7 lines

You can now specify a file in the "addresses.txt" format on the
command-line.

Once the new make-cfg input format is being read in (for jumps), we
will allow multiple such files to be specified.