Don't remove .d files under make clean
------------------------------------------------------------------------
r1973 | jthayer | 2015-02-20 12:50:17 -0600 (Fri, 20 Feb 2015) | 12 lines

splits clean target into clean-deps and clean, where clean deals with
everything that isn't a .d file, and clean-deps removes all of the old
.d files.

Cuts down time between cleans, and the time needed to build after a
clean.  Introduces the minimal risk that we'll accidentally ship some
baked-in d files when we package up the system.

We should really add a dist-clean target that cleans up the whole
directory for tarring and distribution some day.