Update README.txt with credits and a GCC build note
Key to decode SVN usernames to identities, like the one on the
fuzzball cgc-branch with J.'s home page updated.

Add a note about something tricky about how I had to compile GCC on
the UMN machines. Probably not relevant outside UMN CS, but worth
making a note about because it's a non-obvious problem and fix.