Update support for modern Vine IR in static analysis
Parts of this are picked from 8e10f862 and 73834fd9 on the CGC branch.

But was was definitely needed for the examples I was working with
is support for the SLT and SLE binops. The code used to only support
the unsigned LT and LE variants, but since the strided interval
domain uses signed bounds, I believe that implementation was really
more appropriate for SLT and SLE (the signed versions). Leave the old
implementation in place for LT and LE because we don't have anything
better at the moment.