Document and partially clean up a namespace/header file interaction
A problem related to this was causing compilation errors back when
other dependencies were missing, which lead me to investigate why it
was happening. I found out later it was related to a
previously-unresolved TODO issue. This is an approach that does not
require changing LibASMIR itself, since IMO the approach of putting
the LibASMIR includes inside a namespace does not need to be
officially supported.