Add a script for building pre-reqs on UMN CS machines
So far just tested on Ubuntu 16.04 (e.g., ohare.cs). This is a first
step towards codifying everything that's needed to get this code to
compile. Our machines are a hard case because none of the 32-bit
libraries are available as packages, but an easy case because this
script can refer to a bunch of things that are already compiled.